fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	int i = 1, j = 2, k, l = 3;
+	long int i = 1, j = 2, k, l = 3;
 
 	printf("%d, %d", i, j);
 	while (l <= 50)
